move all functional tests into the same folder
